requirements.txt
astroid==2.9.3
attrs==21.4.0
black==22.1.0
click==8.0.3
emoji==1.6.3
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
py==1.11.0
pylint==2.12.2
pyparsing==3.0.7
pytest==6.2.5
python-dotenv==0.19.2
slack-bolt==1.11.4
slack-sdk==3.13.0
toml==0.10.2
tomli==2.0.0
typing-extensions==4.0.1
wrapt==1.13.3
